Simplify CFN app.py file, just use one file, no need for a separate module Remove some redundant comments
