+++ app.py
@@ -1,28 +1,56 @@
 #!/usr/bin/env python3
-import os
+from aws_cdk import (
+    App,
+    Stack,
+    aws_apigateway,
+    aws_lambda,
+    aws_dynamodb,
+)
+from constructs import Construct
 
-import aws_cdk as cdk
 
-from py_rest_api.py_rest_api_stack import PyRestApiStack
+class PyRestApiStack(Stack):
+    def __init__(self, scope: Construct, construct_id: str, **kwargs) -> None:
+        super().__init__(scope, construct_id, **kwargs)
+        todo_table = aws_dynamodb.TableV2(
+            self,
+            "TodoTablePy",
+            partition_key=aws_dynamodb.Attribute(
+                name="id",type=aws_dynamodb.AttributeType.STRING
+            ),
+            billing=aws_dynamodb.Billing.on_demand(),
+        )
+        todo_lambda = aws_lambda.Function(
+            self,
+            "TodoAPILambda",
+            runtime=aws_lambda.Runtime.PYTHON_3_11,
+            code=aws_lambda.Code.from_asset("services"),
+            handler="index.handler",
+            environment={"TABLE_NAME": todo_table.table_name},
+        )
+
+        todo_table.grant_read_write_data(todo_lambda)
+
+        api = aws_apigateway.RestApi(self, "Todo-API-Python")
+        cors_options = aws_apigateway.CorsOptions(
+            allow_origins=aws_apigateway.Cors.ALL_ORIGINS,
+            allow_methods=aws_apigateway.Cors.ALL_METHODS,
+        )
+        todo_resource = api.root.add_resource(
+            "api", default_cors_preflight_options=cors_options
+        )
+
+        todo_lambda_integration = aws_apigateway.LambdaIntegration(todo_lambda)
+        todo_resource.add_method("GET", todo_lambda_integration)
+        todo_resource.add_method("POST", todo_lambda_integration)
 
 
-app = cdk.App()
+app = App()
 PyRestApiStack(app, "TodoAPIPythonStack",
     # If you don't specify 'env', this stack will be environment-agnostic.
     # Account/Region-dependent features and context lookups will not work,
     # but a single synthesized template can be deployed anywhere.
-
-    # Uncomment the next line to specialize this stack for the AWS Account
-    # and Region that are implied by the current CLI configuration.
-
-    #env=cdk.Environment(account=os.getenv('CDK_DEFAULT_ACCOUNT'), region=os.getenv('CDK_DEFAULT_REGION')),
-
-    # Uncomment the next line if you know exactly what Account and Region you
-    # want to deploy the stack to. */
-
-    #env=cdk.Environment(account='123456789012', region='us-east-1'),
-
     # For more information, see https://docs.aws.amazon.com/cdk/latest/guide/environments.html
-    )
+)
 
 app.synth()
