--- py_rest_api/py_rest_api_stack.py
@@ -1,44 +1,0 @@
-from aws_cdk import (
-    Stack,
-    aws_apigateway,
-    aws_lambda,
-    aws_dynamodb,
-)
-from constructs import Construct
-
-
-class PyRestApiStack(Stack):
-
-    def __init__(self, scope: Construct, construct_id: str, **kwargs) -> None:
-        super().__init__(scope, construct_id, **kwargs)
-        todo_table = aws_dynamodb.TableV2(
-            self,
-            "TodoTablePy",
-            partition_key=aws_dynamodb.Attribute(
-                name="id",type=aws_dynamodb.AttributeType.STRING
-            ),
-            billing=aws_dynamodb.Billing.on_demand(),
-        )
-        todo_lambda = aws_lambda.Function(
-            self,
-            "TodoAPILambda",
-            runtime=aws_lambda.Runtime.PYTHON_3_11,
-            code=aws_lambda.Code.from_asset("services"),
-            handler="index.handler",
-            environment={"TABLE_NAME": todo_table.table_name},
-        )
-
-        todo_table.grant_read_write_data(todo_lambda)
-
-        api = aws_apigateway.RestApi(self, "Todo-API-Python")
-        cors_options = aws_apigateway.CorsOptions(
-            allow_origins=aws_apigateway.Cors.ALL_ORIGINS,
-            allow_methods=aws_apigateway.Cors.ALL_METHODS,
-        )
-        todo_resource = api.root.add_resource(
-            "api", default_cors_preflight_options=cors_options
-        )
-
-        todo_lambda_integration = aws_apigateway.LambdaIntegration(todo_lambda)
-        todo_resource.add_method("GET", todo_lambda_integration)
-        todo_resource.add_method("POST", todo_lambda_integration)
